+++ HISTORY.md
@@ -1,11 +1,12 @@
 # cloudpathlib Changelog
 
-## Unreleased
+## v0.21.0 (2025-03-03)
 
+- Removed support for deprecated env var that had a typo (`CLOUPATHLIB_FILE_CACHE_MODE`; you should use `CLOUDPATHLIB_FILE_CACHE_MODE`).
 - Added support for `timeout` and `retry` kwargs for `GSClient`. (Issue [#484](https://github.com/drivendataorg/cloudpathlib/issues/484), PR [#485](https://github.com/drivendataorg/cloudpathlib/pull/485), thanks @Mchristos)
 - Fixed `CloudPath(...) / other` to correctly attempt to fall back on `other`'s `__rtruediv__` implementation, in order to support classes that explicitly support the `/` with a `CloudPath` instance. Previously, this would always raise a `TypeError` if `other` were not a `str` or `PurePosixPath`. (PR [#479](https://github.com/drivendataorg/cloudpathlib/pull/479))
-- Add `md5` property to `GSPath`, updated LocalGSPath to include `md5` property, updated mock_gs.MockBlob to include `md5_hash` property.
-- Fixed an uncaught exception on Azure Gen2 storage accounts with HNS enabled when used with `DefaultAzureCredential`. (Issue [#486](https://github.com/drivendataorg/cloudpathlib/issues/486))
+- Added `md5` property to `GSPath`, updated LocalGSPath to include `md5` property, updated mock_gs.MockBlob to include `md5_hash` property. (Issue [#482](https://github.com/drivendataorg/cloudpathlib/issues/482), PR [#483](https://github.com/drivendataorg/cloudpathlib/pull/453))
+- Fixed an uncaught exception on Azure Gen2 storage accounts with HNS enabled when used with `DefaultAzureCredential`. (Issue [#486](https://github.com/drivendataorg/cloudpathlib/issues/486), PR [#487](https://github.com/drivendataorg/cloudpathlib/pull/487))
 - Removed support for Python 3.8, which was EOL in October 2024. (PR [#502](https://github.com/drivendataorg/cloudpathlib/pull/502))
 
 ## v0.20.0 (2024-10-18)